+++ UsefulWebApps/Repository/NoteSharesRepository.cs
@@ -9,7 +9,7 @@ namespace UsefulWebApps.Repository
     {
         public NoteSharesRepository(MySqlConnection connection) : base(connection) { }
         //any NoteShares model specific database methods here
-        public async Task<bool> ShareNote(int noteId, string sharedWithUserId)
+        public async Task<bool> ShareNote(long noteId, string sharedWithUserId)
         {
             //IGNORE so it will not throw duplicate key error if trying to share same note with same user twice
             string sql = @"INSERT IGNORE INTO note_shares (NoteId, SharedWithUserId) 
@@ -18,7 +18,7 @@ namespace UsefulWebApps.Repository
             return rows > 0;
         }
 
-        public async Task<bool> UnshareNote(int noteId)
+        public async Task<bool> UnshareNote(long noteId)
         {
             string sql = "DELETE FROM note_shares WHERE NoteId = @noteId";
             int rows = await _connection.ExecuteAsync(sql, new { noteId });
@@ -34,7 +34,7 @@ namespace UsefulWebApps.Repository
             return notes;
         }
         // Key: NoteId, Value: list of display names the note is shared with
-        public async Task<Dictionary<int, List<string>>> GetSharedToMapForOwner(string ownerUserId)
+        public async Task<Dictionary<long, List<string>>> GetSharedToMapForOwner(string ownerUserId)
         {
             // For all notes owned by this user that have been shared, return noteId -> friend display names
             // (NoteId, DisplayName of the user the note is shared with)
@@ -44,9 +44,9 @@ namespace UsefulWebApps.Repository
                            INNER JOIN user_profiles up ON up.UserId = ns.SharedWithUserId
                            WHERE n.UserId = @ownerUserId";
 
-            var rows = await _connection.QueryAsync<(int NoteId, string DisplayName)>(sql, new { ownerUserId });
+            var rows = await _connection.QueryAsync<(long NoteId, string DisplayName)>(sql, new { ownerUserId });
 
-            Dictionary<int, List<string>> map = new();
+            Dictionary<long, List<string>> map = new();
             foreach (var row in rows)
             {
                 //if map does not contain key create it
